+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1 (March 31, 2020)
+* html comments inside literal block
+* highlighting decoration disabled inside literal block
+
 ## 1.0.1 (March 29, 2020)
 * formatter dependency fixes and improvements